@@ -12,15 +10,14 @@ plugins {
 apply(plugin = "com.google.gms.google-services")
 
 android {
-    compileSdkVersion(28)
-    buildToolsVersion = "28.0.3"
+    compileSdkVersion(31)
 
     defaultConfig {
         applicationId = "hr.bpervan.novaeva.main"
-        minSdkVersion(16)
-        targetSdkVersion(28)
-        versionCode = 160030100
-        versionName = "3.1.0"
+        minSdkVersion(21)
+        targetSdkVersion(31)
+        versionCode = 210030201
+        versionName = "3.2.1"
     }
 
     buildTypes {
