@@ -4,10 +4,13 @@ import android.content.res.Configuration
 import android.os.Bundle
 import android.os.Handler
 import android.util.Log
+import android.view.View
+import androidx.appcompat.widget.SearchView
 import androidx.core.content.edit
 import androidx.core.os.postDelayed
 import androidx.recyclerview.widget.RecyclerView
-import hr.bpervan.novaeva.NovaEvaApp
+import hr.bpervan.novaeva.EventPipelines
+import hr.bpervan.novaeva.model.CategoryDto
 import hr.bpervan.novaeva.model.EvaDirectory
 import hr.bpervan.novaeva.model.EvaDomainInfo
 import hr.bpervan.novaeva.model.EvaNode
@@ -15,19 +18,24 @@ import hr.bpervan.novaeva.rest.EvaDomain
 import hr.bpervan.novaeva.rest.NovaEvaService
 import hr.bpervan.novaeva.storage.EvaDirectoryDbAdapter
 import hr.bpervan.novaeva.storage.RealmConfigProvider
-import hr.bpervan.novaeva.util.*
+import hr.bpervan.novaeva.util.HAS_NEW_CONTENT_KEY_PREFIX
+import hr.bpervan.novaeva.util.LAST_EVICTION_TIME_MILLIS_KEY_PREFIX
+import hr.bpervan.novaeva.util.dataErrorSnackbar
+import hr.bpervan.novaeva.util.evictionIntervalMillis
+import hr.bpervan.novaeva.util.networkConnectionExists
+import hr.bpervan.novaeva.util.plusAssign
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.Disposable
 import io.reactivex.rxkotlin.subscribeBy
 import io.reactivex.schedulers.Schedulers
 import io.realm.Realm
 import io.realm.kotlin.where
+import java.util.concurrent.TimeUnit
 
 /**
  *
  */
 abstract class EvaAbstractDirectoryFragment : EvaBaseFragment() {
-
 
     private val handler = Handler()
 
@@ -45,6 +53,7 @@ abstract class EvaAbstractDirectoryFragment : EvaBaseFragment() {
     protected var directoryTitle: String = ""
 
     protected var fetchItems: Long = 20
+    protected var searchFetchItems: Long = 20
 
     protected lateinit var adapter: RecyclerView.Adapter<*>
 
@@ -55,6 +64,9 @@ abstract class EvaAbstractDirectoryFragment : EvaBaseFragment() {
     protected var hasMore = true
     protected var fetchingFromServer = true
     protected var loadingFromDb = true
+
+    @Volatile
+    protected var useLocalDb: Boolean = true
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -128,9 +140,17 @@ abstract class EvaAbstractDirectoryFragment : EvaBaseFragment() {
         }
     }
 
+    protected fun unsubscribeFromDirectoryUpdates() {
+        evaDirectoryChangesDisposable?.dispose()
+        elementsList.clear()
+        adapter.notifyDataSetChanged()
+    }
+
+    abstract fun fillElements(categoryDto: CategoryDto)
+
     abstract fun fillElements(evaDirectory: EvaDirectory)
 
-    override fun onConfigurationChanged(newConfig: Configuration?) {
+    override fun onConfigurationChanged(newConfig: Configuration) {
         //A HACK TO DISPLAY CORRECT FRAGMENT VIEWS WHEN SWITCHING BETWEEN PORTRAIT AND LANDSCAPE
         activity?.supportFragmentManager?.beginTransaction()?.detach(this)?.commitAllowingStateLoss()
         super.onConfigurationChanged(newConfig)
@@ -152,6 +172,7 @@ abstract class EvaAbstractDirectoryFragment : EvaBaseFragment() {
 
     protected var pageOn: Long = -99
 
+    @Deprecated("legacy")
     protected fun fetchEvaDirectoryDataFromServer_legacy(timestamp: Long? = null) {
         fetchingFromServer = true
         refreshLoadingCircleState()
@@ -193,25 +214,81 @@ abstract class EvaAbstractDirectoryFragment : EvaBaseFragment() {
                 })
     }
 
-    protected fun fetchEvaDirectoryDataFromServer(page: Long = 1) {
+    protected fun initSearch(searchView: SearchView) {
+
+        disposables += EventPipelines.search
+            .debounce(500, TimeUnit.MILLISECONDS)
+            .observeOn(AndroidSchedulers.mainThread())
+            .subscribe { searchQuery ->
+                onSearch(searchQuery)
+            }
+
+        if (domain.isLegacy()) {
+            searchView.visibility = View.GONE
+        } else {
+            searchView.setOnQueryTextListener(object : SearchView.OnQueryTextListener {
+
+                override fun onQueryTextSubmit(query: String): Boolean {
+                    onSearch(query.trim())
+                    return true
+                }
+
+                override fun onQueryTextChange(newText: String): Boolean {
+                    EventPipelines.search.onNext(newText.trim())
+                    return true
+                }
+            })
+        }
+    }
+
+    protected fun onSearch(searchQuery: String) {
+        if (useLocalDb && searchQuery.length >= 3) {
+            useLocalDb = false
+            unsubscribeFromDirectoryUpdates() // will clear elements
+        } else if (!useLocalDb && searchQuery.length < 3) {
+            useLocalDb = true
+            subscribeToDirectoryUpdates() // will fill elements
+        }
+
+        if (!useLocalDb) {
+            fetchEvaDirectoryDataFromServer(searchQuery = searchQuery)
+        }
+    }
+
+    protected fun fetchEvaDirectoryDataFromServer(page: Long = 1, searchQuery: String? = null) {
         fetchingFromServer = true
         refreshLoadingCircleState()
 
         fetchFromServerDisposable = NovaEvaService.v3
-                .categoryContent(
+                    .categoryContent(
                         domain = domain.domainEndpoint,
                         categoryId = directoryId,
                         page = page,
-                        items = fetchItems)
-                .subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribeBy(onSuccess = { categoryDto ->
-                    loadingFromDb = true
+                        items = if (searchQuery == null) fetchItems else searchFetchItems,
+                        searchQuery = searchQuery
+                    )
+                    .subscribeOn(Schedulers.io())
+                    .observeOn(AndroidSchedulers.mainThread())
+                    .subscribeBy(onSuccess = { categoryDto ->
+
                     fetchingFromServer = false
+
+                    if (!useLocalDb) {
+                        loadingFromDb = false
+                        fillElements(categoryDto)
+                        adapter.notifyDataSetChanged()
+                        return@subscribeBy
+                    }
+
+                    loadingFromDb = true
 
                     if (directoryId == 0L) {
                         directoryId = categoryDto.id
                         createIfMissingAndSubscribeToEvaDirectoryUpdates()
+                    }
+
+                    if (categoryDto.domain == null) {
+                        categoryDto.domain = domain
                     }
 
                     EvaDirectoryDbAdapter.addOrUpdateEvaCategoryAsync(realm, categoryDto)
