@@ -1,24 +1,20 @@
 package hr.bpervan.novaeva.fragments
 
-import android.app.AlertDialog
 import android.content.Context
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.EditText
 import com.google.firebase.analytics.FirebaseAnalytics
-import hr.bpervan.novaeva.NovaEvaApp
 import hr.bpervan.novaeva.EventPipelines
 import hr.bpervan.novaeva.adapters.EvaRecyclerAdapter
 import hr.bpervan.novaeva.main.R
+import hr.bpervan.novaeva.main.databinding.FragmentBookmarksBinding
 import hr.bpervan.novaeva.model.EvaContent
 import hr.bpervan.novaeva.storage.EvaContentDbAdapter
 import hr.bpervan.novaeva.storage.RealmConfigProvider
 import io.reactivex.disposables.Disposable
 import io.realm.Realm
-import kotlinx.android.synthetic.main.fragment_bookmarks.*
-import java.util.*
 
 /**
  *
@@ -32,6 +28,8 @@ class EvaBookmarksFragment : EvaBaseFragment() {
         }
     }
 
+    private var _viewBinding: FragmentBookmarksBinding? = null
+    private val viewBinding get() = _viewBinding!!
 
     private var loadBookmarksFromDbDisposable: Disposable? = null
         set(value) {
@@ -52,7 +50,8 @@ class EvaBookmarksFragment : EvaBaseFragment() {
     }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.fragment_bookmarks, container, false)
+        _viewBinding = FragmentBookmarksBinding.inflate(inflater, container, false)
+        return viewBinding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -62,14 +61,14 @@ class EvaBookmarksFragment : EvaBaseFragment() {
         EventPipelines.changeStatusbarColor.onNext(R.color.VeryDarkGray)
         EventPipelines.changeFragmentBackgroundResource.onNext(R.color.White)
 
-        val recyclerView = evaRecyclerView as androidx.recyclerview.widget.RecyclerView
+        val recyclerView = viewBinding.evaRecyclerView.root
 
         recyclerView.adapter = adapter
         recyclerView.layoutManager = androidx.recyclerview.widget.LinearLayoutManager(recyclerView.context)
         recyclerView.itemAnimator = androidx.recyclerview.widget.DefaultItemAnimator()
     }
 
-    override fun onAttach(context: Context?) {
+    override fun onAttach(context: Context) {
         super.onAttach(context)
 
         activity?.title = "Bookmarks"
@@ -85,6 +84,11 @@ class EvaBookmarksFragment : EvaBaseFragment() {
                 .setCurrentScreen(requireActivity(), "Zabilješke", "Bookmarks")
     }
 
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _viewBinding = null
+    }
+
     private fun reloadBookmarksFromDb() {
         bookmarksList.clear()
         loadBookmarksFromDbDisposable =
@@ -92,22 +96,5 @@ class EvaBookmarksFragment : EvaBaseFragment() {
                     bookmarksList.add(it)
                     adapter.notifyDataSetChanged()
                 }
-    }
-
-
-    private fun showSearchPopup() {
-        activity?.let { activity ->
-            val searchBuilder = AlertDialog.Builder(activity)
-            searchBuilder.setTitle("Pretraga")
-            val et = EditText(activity)
-            searchBuilder.setView(et)
-            searchBuilder.setPositiveButton("Pretrazi") { _, _ ->
-                val searchString = et.text.toString()
-                EventPipelines.search.onNext(searchString)
-            }
-            searchBuilder.setNegativeButton("Odustani") { _, _ -> }
-
-            searchBuilder.show()
-        }
     }
 }
