@@ -12,11 +12,12 @@ import hr.bpervan.novaeva.EventPipelines
 import hr.bpervan.novaeva.NovaEvaApp
 import hr.bpervan.novaeva.adapters.PrayerBookRecyclerAdapter
 import hr.bpervan.novaeva.main.R
+import hr.bpervan.novaeva.main.databinding.FragmentPrayersBinding
+import hr.bpervan.novaeva.model.CategoryDto
 import hr.bpervan.novaeva.model.EvaDirectory
 import hr.bpervan.novaeva.model.OpenPrayerDirectoryEvent
+import hr.bpervan.novaeva.model.toDbModel
 import hr.bpervan.novaeva.rest.EvaDomain
-import kotlinx.android.synthetic.main.fragment_prayers.*
-import kotlinx.android.synthetic.main.top_prayerbook.*
 
 /**
  * Created by vpriscan on 11.12.17..
@@ -31,11 +32,14 @@ class PrayerBookFragment : EvaAbstractDirectoryFragment() {
         }
     }
 
+    private var _viewBinding: FragmentPrayersBinding? = null
+    private val viewBinding get() = _viewBinding!!
+
     private lateinit var initializer: OpenPrayerDirectoryEvent
 
     override fun onCreate(savedInstanceState: Bundle?) {
 
-        val inState: Bundle = savedInstanceState ?: arguments!!
+        val inState: Bundle = savedInstanceState ?: requireArguments()
 
         initializer = inState.getParcelable(EvaFragmentFactory.INITIALIZER)!!
 
@@ -43,7 +47,8 @@ class PrayerBookFragment : EvaAbstractDirectoryFragment() {
         directoryId = initializer.directoryId
         directoryTitle = initializer.title
 
-        fetchItems = 1000
+        fetchItems = 500
+        searchFetchItems = 20
 
         adapter = PrayerBookRecyclerAdapter(elementsList)
 
@@ -55,9 +60,10 @@ class PrayerBookFragment : EvaAbstractDirectoryFragment() {
         super.onSaveInstanceState(outState)
     }
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater.cloneInContext(ContextThemeWrapper(activity, R.style.PrayersTheme))
-                .inflate(R.layout.fragment_prayers, container, false)
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
+        val newInflater = inflater.cloneInContext(ContextThemeWrapper(activity, R.style.PrayersTheme))
+        _viewBinding = FragmentPrayersBinding.inflate(newInflater, container, false)
+        return viewBinding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -67,17 +73,26 @@ class PrayerBookFragment : EvaAbstractDirectoryFragment() {
         EventPipelines.changeStatusbarColor.onNext(R.color.VeryDarkGray)
         EventPipelines.changeFragmentBackgroundResource.onNext(R.color.White)
 
-        prayerArrow.isInvisible = true
-        prayerTitleTextView.apply {
+        viewBinding.collapsingPrayerHeader.prayerTop.prayerArrow.isInvisible = true
+        viewBinding.collapsingPrayerHeader.prayerTop.prayerTitleTextView.apply {
             text = directoryTitle
             typeface = NovaEvaApp.openSansBold
         }
 
-        val recyclerView = evaRecyclerView as androidx.recyclerview.widget.RecyclerView
+        val recyclerView = viewBinding.evaRecyclerView.root
         val linearLayoutManager = androidx.recyclerview.widget.LinearLayoutManager(context)
         recyclerView.layoutManager = linearLayoutManager
         recyclerView.itemAnimator = androidx.recyclerview.widget.DefaultItemAnimator()
         recyclerView.adapter = adapter
+
+        val coverImageView = viewBinding.collapsingPrayerHeader.prayerCoverImage
+        val url = prefs.getString("hr.bpervan.novaeva.categoryheader.$domain", null)
+        if (url != null) {
+            imageLoader.displayImage(url, coverImageView)
+        }
+
+        val searchView = viewBinding.collapsingPrayerHeader.prayerTop.prayerbookSearchView
+        initSearch(searchView)
     }
 
     override fun onResume() {
@@ -85,6 +100,20 @@ class PrayerBookFragment : EvaAbstractDirectoryFragment() {
 
         FirebaseAnalytics.getInstance(requireContext())
                 .setCurrentScreen(requireActivity(), "Molitvenik", "PrayerBook")
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _viewBinding = null
+    }
+
+    override fun fillElements(categoryDto: CategoryDto) {
+
+        elementsList.clear()
+
+        val subcategoriesSorted = categoryDto.subcategories.orEmpty().map{it.toDbModel()}.sortedByDescending { it.position }
+
+        elementsList.addAll(subcategoriesSorted)
     }
 
     override fun fillElements(evaDirectory: EvaDirectory) {
