@@ -4,19 +4,15 @@ import android.util.Log
 import androidx.annotation.StringRes
 import hr.bpervan.novaeva.main.BuildConfig
 import hr.bpervan.novaeva.main.R
-import hr.bpervan.novaeva.rest.EvaDomain.VOCATION
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import java.security.cert.X509Certificate
 import javax.net.ssl.SSLContext
 import javax.net.ssl.X509TrustManager
 
-/**
- * Todo multiple regions
- */
 object Region {
     const val name: String = "Croatia"
-    const val id: Long = 0
+    const val id: Long = 1
 }
 
 enum class Server(val baseUrl: String, val auth: String?) {
@@ -36,19 +32,14 @@ enum class EvaDomain(@StringRes val title: Int,
     SERMONS(R.string.sermons, "sermons"),
     VOCATION(R.string.vocation, "vocation"),
     ANSWERS(R.string.answers, "answers"),
-    SONGBOOK(R.string.songbook, "songbook", 355),
+    SONGS(R.string.songbook, "songs"),
     RADIO(R.string.radio, "radio"),
-    PRAYERS(R.string.prayerbook, "prayers");
+    PRAYERS(R.string.prayerbook, "prayers"),
+    LIVE(R.string.live, "live")
+    ;
 
     fun isLegacy(): Boolean {
-        return this == EvaDomain.GOSPEL || this == EvaDomain.SONGBOOK
-    }
-}
-
-fun serverByDomain(domain: EvaDomain): Server {
-    return when (domain) {
-        VOCATION -> Server.V3
-        else -> Server.V2
+        return this == EvaDomain.GOSPEL
     }
 }
 
