@@ -10,30 +10,30 @@ import android.util.Log
  */
 class LifecycleLogger : Application.ActivityLifecycleCallbacks {
     override fun onActivityCreated(activity: Activity, savedInstanceState: Bundle?) {
-        Log.d("activityLifecycle", "Activity created:$activity")
+        Log.d("activityLifecycle", "Activity created: $activity")
     }
 
     override fun onActivityStarted(activity: Activity) {
-        Log.d("activityLifecycle", "Activity started:$activity")
+        Log.d("activityLifecycle", "Activity started: $activity")
     }
 
     override fun onActivityResumed(activity: Activity) {
-        Log.d("activityLifecycle", "Activity resumed:$activity")
+        Log.d("activityLifecycle", "Activity resumed: $activity")
     }
 
     override fun onActivityPaused(activity: Activity) {
-        Log.d("activityLifecycle", "Activity paused:$activity")
+        Log.d("activityLifecycle", "Activity paused: $activity")
     }
 
     override fun onActivityStopped(activity: Activity) {
-        Log.d("activityLifecycle", "Activity stopped:$activity")
+        Log.d("activityLifecycle", "Activity stopped: $activity")
     }
 
-    override fun onActivitySaveInstanceState(activity: Activity, outState: Bundle?) {
-        Log.d("activityLifecycle", "Activity saved instance state:$activity")
+    override fun onActivitySaveInstanceState(activity: Activity, outState: Bundle) {
+        Log.d("activityLifecycle", "Activity saved instance state: $activity")
     }
 
     override fun onActivityDestroyed(activity: Activity) {
-        Log.d("activityLifecycle", "Activity destroyed:$activity")
+        Log.d("activityLifecycle", "Activity destroyed: $activity")
     }
 }
